+++ core/src/test/java/io/confluent/rest/SaslTest.java
@@ -16,8 +16,6 @@
 
 package io.confluent.rest;
 
-import io.confluent.common.metrics.KafkaMetric;
-import io.confluent.rest.annotations.PerformanceMetric;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
@@ -30,6 +28,15 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.StandardOpenOption;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Map;
+import java.util.Properties;
+
 import javax.security.auth.login.Configuration;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -41,16 +48,13 @@ import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
-import java.io.File;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.StandardOpenOption;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Map;
-import java.util.Properties;
 
-import static org.junit.Assert.*;
+import io.confluent.common.metrics.KafkaMetric;
+import io.confluent.rest.annotations.PerformanceMetric;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertTrue;
 
 public class SaslTest {
   private static final Logger log = LoggerFactory.getLogger(SaslTest.class);
