+++ core/src/test/java/io/confluent/rest/StaticResourcesTest.java
@@ -23,16 +23,17 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+import java.util.Properties;
+
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.core.Configurable;
 import javax.ws.rs.core.Response;
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.Properties;
 
 import static org.junit.Assert.assertEquals;
 
@@ -67,12 +68,12 @@ public class StaticResourcesTest {
 
   @Test
   public void testStaticContent() throws Exception {
-    testGet("/index.html", 200, staticContent.trim());
+    testGet("/index.html", 200, staticContent);
   }
 
   @Test
   public void testDefaultServletMapsToIndex() throws Exception {
-    testGet("/", 200, staticContent.trim());
+    testGet("/", 200, staticContent);
   }
 
   @Test
@@ -87,8 +88,7 @@ public class StaticResourcesTest {
         .request()
         .get();
     assertEquals(expectedStatus, response.getStatus());
-    final String entity = response.readEntity(String.class);
-    assertEquals(expectedMessage, entity == null ? null : entity.trim());
+    assertEquals(expectedMessage, response.readEntity(String.class));
   }
 
   private static class StaticApplication extends Application<TestRestConfig> {
