--- examples/src/test/java/io/confluent/rest/examples/IdleTimeoutConfigTest.java
@@ -1,46 +1,0 @@
-package io.confluent.rest.examples;
-
-import io.confluent.rest.RestConfig;
-import org.eclipse.jetty.server.Server;
-import org.junit.Assert;
-import org.junit.Test;
-
-import javax.ws.rs.core.Configurable;
-import java.util.HashMap;
-import java.util.Map;
-
-
-public class IdleTimeoutConfigTest {
-
-    @Test
-    public void testIdleTimeoutConfigIsApplied() throws Exception {
-
-        // given
-        long expectedIdleTimeout = 1000;
-
-        // when
-        Map<String, Object> props = new HashMap<>();
-        props.put(RestConfig.IDLE_TIMEOUT_MS_CONFIG, String.valueOf(expectedIdleTimeout));
-        RestConfig config = new RestConfig(RestConfig.baseConfigDef(), props);
-
-        Server server = new TestApp(config).createServer();
-
-        // then
-        Assert.assertEquals(expectedIdleTimeout, server.getConnectors()[0].getIdleTimeout());
-
-    }
-
-
-    public static class TestApp extends io.confluent.rest.Application<RestConfig> {
-
-        public TestApp(final RestConfig config) {
-            super(config);
-        }
-
-
-        @Override
-        public void setupResources(final Configurable<?> config, final RestConfig appConfig) {
-            // nothing - not needed
-        }
-    }
-}
